Create FoundationExtensions
Merge common codes to FoundationExtensions
