+++ Package.swift
@@ -4,7 +4,11 @@ let package = Package(
     name: "DevHelperToolkit",
     targets: [
         Target(name: "DevHelperToolkit",
-               dependencies: ["JSONHelperToolkit", "ColorHelperToolkit", "ImageHelperToolkit", "StoryboardHelperToolkit"])
+               dependencies: ["JSONHelperToolkit", "ColorHelperToolkit", "ImageHelperToolkit", "StoryboardHelperToolkit"]),
+        Target(name: "JSONHelperToolkit", dependencies: ["FoundationExtensions"]),
+        Target(name: "ImageHelperToolkit", dependencies: ["FoundationExtensions"]),
+        Target(name: "StoryboardHelperToolkit", dependencies: ["FoundationExtensions"]),
+        Target(name: "FoundationExtensions")
     ],
     exclude: ["Resources"]
 )
